feat: implement CRUDs and update outputs for frontend needs
- Updated multiple output DTOs to align with frontend requirements
- Added missing CRUD functionalities for various modules
- Improved permission handling for groups and users
- Renamed and refactored repositories, DTOs, and controllers for better organization
- Adjusted migrations and seeders to reflect new permissions structure
- Refactored routes and middleware for enhanced permission checks

This commit addresses feature requests from the frontend to improve integration and system functionality.

+++ app/Application/Stock/Supplier/DeleteSupplier.php
@@ -3,15 +3,18 @@
 namespace App\Application\Stock\Supplier;
 
 use App\Application\DTO\IdRequestDTO;
+use App\Domain\Exceptions\Stock\ProductCategory\ProductCategoryException;
 use App\Domain\Exceptions\Stock\Supplier\SupplierException;
 use App\Domain\Exceptions\Stock\Supplier\SupplierNotFoundException;
+use App\Domain\Interfaces\Stock\Product\ProductRepositoryInterface;
 use App\Domain\Interfaces\Stock\Supplier\SupplierRepositoryInterface;
 
 class DeleteSupplier
 {
 
     public function __construct(
-        private SupplierRepositoryInterface $supplierRepositoryInterface
+        private SupplierRepositoryInterface $supplierRepositoryInterface,
+        private ProductRepositoryInterface $productRepositoryInterface
     ) {}
 
     public function execute(IdRequestDTO $input): bool
@@ -19,6 +22,10 @@ class DeleteSupplier
         $supplier = $this->supplierRepositoryInterface->getById($input->getId());
         if (empty($supplier)) {
             throw new SupplierNotFoundException('Fornecedor não encontrado', 404);
+        }
+
+        if(!empty($this->productRepositoryInterface->getBySupplierId($supplier->id))) {
+            throw new ProductCategoryException('Fornecedor não pode ser deletado, pois existem produtos vinculados a ele', 400);
         }
 
         if (!$this->supplierRepositoryInterface->delete($supplier)) {
